Revert upload summary
This PR reverts some changes from #9. Since we had to restrict downloads.rapids.ai to be VPN-accessible only, it doesn't make sense to broadcast the URL so prominently anymore. Also, it was a bit cluttersome that the same summary card was being created for every matrix variant in our builds.