Generate stable version for pip wheel release builds (#33)
Currently in the pip wheel workflows, the flow is as follows:
* If the user does not input their desired stable version string (e.g.
`23.02.00`) as a workflow input, the script `rapids-pip-wheel-version`
generates it automatically with the following logic:
1. In the workflow, the epoch timestamp of the current build is
generated and passed as the argument: `rapids-pip-wheel-version
$epoch_timestamp`
    2. Script gets the tag via git describe tag --abbrev0: `v23.02.00a`
    3. Script removes letters: `23.02.00`
4. Script runs `distutils.normalize` (strips some leading 0s etc. by
python versioning rules): `23.2.0`
    5. Script appends the epoch timestamp: `23.2.0.$epoch_timestamp`
* If the user **does** input their desired stable version string as a
workflow input, that string is used directly as the wheel version.
However, in practice, the user always inputs the same thing that step 4
above generates from the git tag
* Therefore, it's much simpler to do the following:
    1. Drop the version override as a workflow input
    2. Embed the logic into the rapids-pip-wheel-version script:
a. If rapids-is-release-build: stop at step 4 (use `23.2.0` as a
version)
b. If **not** rapids-is-release-build: append the epoch timestamp
(`23.2.0.$epoch_timestamp`) as version (same behavior as before)

This PR adds the new behavior described.

This way we eliminate all user inputs and have automatic stable/nightly
versioning generated by the workflow.