Ensure `VERSION_SUFFIX` is added to PR artifacts (#37)
This change ensures that the `VERSION_SUFFIX` environment variable is
available for pull-request builds.

This will ensure that pull-request and nightly packages can be installed
in the same environment.

@bdice had some issues with this recently due to this discrepancy in
pull-request vs. nightly packages versions.

In the long term, we should explore moving the `VERSION_SUFFIX` value to
the build string of our conda packages rather than the version string
since having the value in the version string can cause issues and is
inconsistent with our stable build versions.