Use more robust method to launch twine (#52)
Using the twine executable directly relies on the Python bin directory
being added to the path, which is not reliably the case in different
environments. I've created a new copy of the script since I'm iterating
very actively right now in the new wheels workflow. We can get rid of
the old script once I've verified that everything is working as
expected.