Make `rapids-get-artifact` work locally
This PR ensures that `rapids-get-artifact` works locally.

This request was made by @bdice who was attempting to use the script locally to debug some CUDA 12 issues.

I've moved the logic for downloading, extracting, and printing the artifact's path to a new script, `_rapids-download-from-s3`.

This script is then called by both `rapids-download-from-s3` and `rapids-get-artifact`.

In the future, I would like to rework some of the scripts regarding artifact uploading/downloading, but I think we need to finish reworking the wheel workflows before that will be feasible.